Fix relative links in Jekyll
While GitHub has no issues with relative links between files, Jekyll
does not work with them out of the box. For example, `[title](title.md)`
will create a link to `https://domain.com/title.md`, not
`https://domain.com/title.html`. This will result in a 404 on the GitHub
pages URL. The `jekyll-relative-links` plugin changes the links to point
to the html URLs correctly, leading to a consistent experience between
the Markdown and GitHub pages.

Fixes #137